chore: 封装消息通知组件
